emu: the A9 cross-build claim is now tested, and the notebook consolidated
(B) The Makefile has carried "the same sources are meant to cross-compile for
the A9 later, so nothing here may depend on the host being 64-bit or
little-endian" for a long time with nobody testing it.  arm-none-eabi-gcc is
available, so now it is tested: `make arm` builds the eight core files for a
Cortex-A9 under -Wconversion -Werror, and it runs as part of `make test`.  The
claim holds -- 43.5 KB of ARM text, warning-free.

Two real warnings had to go first: four flag clears where `~XTF_x` promotes to
int before the &=, and `(uint8_t)~p->irqst == 0`, which is correct but reads as
a bug and warns on gcc.  Rewritten as `p->irqst == 0xFF`, which is what it
means.  All twelve gates green afterwards and the ACID score is unmoved.

(A) Consolidated emu/README.md.  Its Status block still claimed 42/63 against a
fabric 33 and carried a duplicated DMA bullet; it now says 48/63 against 32,
notes that the DMA table has no WIDE rows so wide geometry rests solely on
antic_virtdma, records the bench figures, and explains why the five mod_*
modules are counted rather than reclassified.

Added a "How to read this file" section -- it is a chronological notebook, not a
specification, and the headings carry the verdict (SOLVED/CLOSED, Open/PARKED,
SUPERSEDED, Disproved).  Nine headings that described work later overtaken are
now marked SUPERSEDED rather than reading as live, including "Blocked on an OS
ROM: antic_virtdma" for a test that now passes and two pokey_inittiming sections
whose split was dissolved.

No score change: 48 of 63, twelve gates green.

+++ emu/xt6502.c
@@ -506,13 +506,13 @@ void xt6502_step(xt6502 *c)
     case 0xC8: rd(c, c->pc); c->y++; setnz(c, c->y); break;
     case 0xCA: rd(c, c->pc); c->x--; setnz(c, c->x); break;
     case 0x88: rd(c, c->pc); c->y--; setnz(c, c->y); break;
-    case 0x18: rd(c, c->pc); c->p &= ~XTF_C; break;
+    case 0x18: rd(c, c->pc); c->p &= (uint8_t)~XTF_C; break;
     case 0x38: rd(c, c->pc); c->p |=  XTF_C; break;
-    case 0x58: rd(c, c->pc); c->p &= ~XTF_I; break;
+    case 0x58: rd(c, c->pc); c->p &= (uint8_t)~XTF_I; break;
     case 0x78: rd(c, c->pc); c->p |=  XTF_I; break;
-    case 0xD8: rd(c, c->pc); c->p &= ~XTF_D; break;
+    case 0xD8: rd(c, c->pc); c->p &= (uint8_t)~XTF_D; break;
     case 0xF8: rd(c, c->pc); c->p |=  XTF_D; break;
-    case 0xB8: rd(c, c->pc); c->p &= ~XTF_V; break;
+    case 0xB8: rd(c, c->pc); c->p &= (uint8_t)~XTF_V; break;
     case 0xEA: rd(c, c->pc); break;                          /* NOP */
 
     /* ---- stack ---- */
